+++ packages/pigeon/example/app/ios/Runner/Messages.g.swift
@@ -159,6 +159,7 @@ protocol ExampleHostApi {
   func getHostLanguage() throws -> String
   func add(_ a: Int64, to b: Int64) throws -> Int64
   func sendMessage(message: MessageData, completion: @escaping (Result<Bool, Error>) -> Void)
+  func sendMessageModernAsync(message: MessageData) async throws -> Bool
 }
 
 /// Generated setup class from Pigeon to handle messages through the `binaryMessenger`.
@@ -221,6 +222,30 @@ class ExampleHostApiSetup {
       }
     } else {
       sendMessageChannel.setMessageHandler(nil)
+    }
+    let sendMessageModernAsyncChannel = FlutterBasicMessageChannel(
+      name:
+        "dev.flutter.pigeon.pigeon_example_package.ExampleHostApi.sendMessageModernAsync\(channelSuffix)",
+      binaryMessenger: binaryMessenger, codec: codec)
+    if let api = api {
+      sendMessageModernAsyncChannel.setMessageHandler { message, reply in
+        let args = message as! [Any?]
+        let messageArg = args[0] as! MessageData
+        Task {
+          do {
+            let result = try await api.sendMessageModernAsync(message: messageArg)
+            await MainActor.run {
+              reply(wrapResult(result))
+            }
+          } catch {
+            await MainActor.run {
+              reply(wrapError(error))
+            }
+          }
+        }
+      }
+    } else {
+      sendMessageModernAsyncChannel.setMessageHandler(nil)
     }
   }
 }