+++ packages/pigeon/example/app/macos/Runner/messages.g.h
@@ -51,6 +51,9 @@ NSObject<FlutterMessageCodec> *PGNGetMessagesCodec(void);
                            error:(FlutterError *_Nullable *_Nonnull)error;
 - (void)sendMessageMessage:(PGNMessageData *)message
                 completion:(void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion;
+- (void)sendMessageModernAsyncMessage:(PGNMessageData *)message
+                           completion:
+                               (void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion;
 @end
 
 extern void SetUpPGNExampleHostApi(id<FlutterBinaryMessenger> binaryMessenger,