+++ packages/pigeon/example/app/pigeons/messages.dart
@@ -52,6 +52,9 @@ abstract class ExampleHostApi {
 
   @async
   bool sendMessage(MessageData message);
+
+  @modernAsync
+  bool sendMessageModernAsync(MessageData message);
 }
 // #enddocregion host-definitions