+++ packages/pigeon/lib/pigeon_lib.dart
@@ -41,6 +41,37 @@ class _Asynchronous {
   const _Asynchronous();
 }
 
+/// Metadata to annotate a Api method as asynchronous
+const Object async = _Asynchronous();
+
+class _ModernAsynchronous {
+  const _ModernAsynchronous();
+}
+
+/// Provides a modern asynchronous Api (only Swift and Kotlin).
+///
+/// Example:
+///
+///```dart
+///@HostApi()
+///abstract class ExampleHostApi {
+///  @modernAsync
+///  bool sendMessage(MessageData message);
+///}
+///```
+/// Swift(iOS 13.0+):
+///
+/// ```swift
+/// func sendMessage(message: MessageData) async -> Bool
+/// ```
+///
+/// Kotlin (`ExampleHostApi.setUp` will require `CoroutineScope`):
+///
+/// ```kotlin
+/// suspend fun sendMessage(message: MessageData) : Boolean
+/// ```
+const Object modernAsync = _ModernAsynchronous();
+
 class _Attached {
   const _Attached();
 }
@@ -48,9 +79,6 @@ class _Attached {
 class _Static {
   const _Static();
 }
-
-/// Metadata to annotate a Api method as asynchronous
-const Object async = _Asynchronous();
 
 /// Metadata to annotate the field of a ProxyApi as an Attached Field.
 ///
@@ -2002,7 +2030,9 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
     final dart_ast.FormalParameterList parameters = node.parameters!;
     final List<Parameter> arguments =
         parameters.parameters.map(formalParameterToPigeonParameter).toList();
-    final bool isAsynchronous = _hasMetadata(node.metadata, 'async');
+
+    final AsynchronousType asynchronousType =
+        _parseAsynchronousType(node.metadata);
     final bool isStatic = _hasMetadata(node.metadata, 'static');
     final String objcSelector = _findMetadata(node.metadata, 'ObjCSelector')
             ?.arguments
@@ -2050,13 +2080,13 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
             AstFlutterApi() => ApiLocation.flutter,
             AstEventChannelApi() => ApiLocation.host,
           },
-          isAsynchronous: isAsynchronous,
           objcSelector: objcSelector,
           swiftFunction: swiftFunction,
           offset: node.offset,
           taskQueueType: taskQueueType,
           documentationComments:
               _documentationCommentsParser(node.documentationComment?.tokens),
+          asynchronousType: asynchronousType,
         ),
       );
     } else if (_currentClass != null) {
@@ -2244,6 +2274,8 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
       final TaskQueueType taskQueueType =
           _stringToEnum(TaskQueueType.values, taskQueueTypeName) ??
               TaskQueueType.serial;
+      final AsynchronousType asynchronousType =
+          _parseAsynchronousType(node.metadata);
 
       // Methods without named return types aren't supported.
       final dart_ast.TypeAnnotation returnType = type.returnType!;
@@ -2262,7 +2294,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
           isRequired: type.question == null,
           isStatic: isStatic,
           parameters: parameters,
-          isAsynchronous: _hasMetadata(node.metadata, 'async'),
+          asynchronousType: asynchronousType,
           swiftFunction: swiftFunction,
           offset: node.offset,
           taskQueueType: taskQueueType,
@@ -2301,6 +2333,19 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
       }
     }
   }
+}
+
+AsynchronousType _parseAsynchronousType(
+    dart_ast.NodeList<dart_ast.Annotation> metadata) {
+  if (_hasMetadata(metadata, 'async')) {
+    return AsynchronousType.callback;
+  }
+
+  if (_hasMetadata(metadata, 'modernAsync')) {
+    return AsynchronousType.modern;
+  }
+
+  return AsynchronousType.none;
 }
 
 int? _calculateLineNumberNullable(String contents, int? offset) {