+++ packages/pigeon/test/objc_generator_test.dart
@@ -1330,19 +1330,20 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[
-              Parameter(
-                  type: TypeDeclaration(
-                    baseName: 'Input',
-                    associatedClass: emptyClass,
-                    isNullable: false,
-                  ),
-                  name: 'input')
-            ],
-            returnType: const TypeDeclaration.voidDeclaration(),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[
+            Parameter(
+                type: TypeDeclaration(
+                  baseName: 'Input',
+                  associatedClass: emptyClass,
+                  isNullable: false,
+                ),
+                name: 'input')
+          ],
+          returnType: const TypeDeclaration.voidDeclaration(),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[
       Class(name: 'Input', fields: <NamedType>[
@@ -1381,23 +1382,24 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[
-              Parameter(
-                  type: TypeDeclaration(
-                    baseName: 'Input',
-                    associatedClass: emptyClass,
-                    isNullable: false,
-                  ),
-                  name: 'input')
-            ],
-            returnType: TypeDeclaration(
-              baseName: 'Output',
-              associatedClass: emptyClass,
-              isNullable: false,
-            ),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[
+            Parameter(
+                type: TypeDeclaration(
+                  baseName: 'Input',
+                  associatedClass: emptyClass,
+                  isNullable: false,
+                ),
+                name: 'input')
+          ],
+          returnType: TypeDeclaration(
+            baseName: 'Output',
+            associatedClass: emptyClass,
+            isNullable: false,
+          ),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[
       Class(name: 'Input', fields: <NamedType>[
@@ -1436,15 +1438,16 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[],
-            returnType: TypeDeclaration(
-              baseName: 'Output',
-              associatedClass: emptyClass,
-              isNullable: false,
-            ),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[],
+          returnType: TypeDeclaration(
+            baseName: 'Output',
+            associatedClass: emptyClass,
+            isNullable: false,
+          ),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[
       Class(name: 'Output', fields: <NamedType>[
@@ -1478,11 +1481,12 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[],
-            returnType: const TypeDeclaration.voidDeclaration(),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[],
+          returnType: const TypeDeclaration.voidDeclaration(),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[], enums: <Enum>[]);
     final StringBuffer sink = StringBuffer();
@@ -1510,23 +1514,24 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[
-              Parameter(
-                  type: TypeDeclaration(
-                    baseName: 'Input',
-                    associatedClass: emptyClass,
-                    isNullable: false,
-                  ),
-                  name: '')
-            ],
-            returnType: TypeDeclaration(
-              baseName: 'Output',
-              associatedClass: emptyClass,
-              isNullable: false,
-            ),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[
+            Parameter(
+                type: TypeDeclaration(
+                  baseName: 'Input',
+                  associatedClass: emptyClass,
+                  isNullable: false,
+                ),
+                name: '')
+          ],
+          returnType: TypeDeclaration(
+            baseName: 'Output',
+            associatedClass: emptyClass,
+            isNullable: false,
+          ),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[
       Class(name: 'Input', fields: <NamedType>[
@@ -1565,19 +1570,20 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[
-              Parameter(
-                  type: TypeDeclaration(
-                    baseName: 'Input',
-                    associatedClass: emptyClass,
-                    isNullable: false,
-                  ),
-                  name: 'foo')
-            ],
-            returnType: const TypeDeclaration.voidDeclaration(),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[
+            Parameter(
+                type: TypeDeclaration(
+                  baseName: 'Input',
+                  associatedClass: emptyClass,
+                  isNullable: false,
+                ),
+                name: 'foo')
+          ],
+          returnType: const TypeDeclaration.voidDeclaration(),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[
       Class(name: 'Input', fields: <NamedType>[
@@ -1616,11 +1622,12 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[],
-            returnType: const TypeDeclaration.voidDeclaration(),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[],
+          returnType: const TypeDeclaration.voidDeclaration(),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[], enums: <Enum>[]);
     final StringBuffer sink = StringBuffer();
@@ -1648,15 +1655,16 @@ void main() {
     final Root root = Root(apis: <Api>[
       AstHostApi(name: 'Api', methods: <Method>[
         Method(
-            name: 'doSomething',
-            location: ApiLocation.host,
-            parameters: <Parameter>[],
-            returnType: TypeDeclaration(
-              baseName: 'Output',
-              associatedClass: emptyClass,
-              isNullable: false,
-            ),
-            isAsynchronous: true)
+          name: 'doSomething',
+          location: ApiLocation.host,
+          parameters: <Parameter>[],
+          returnType: TypeDeclaration(
+            baseName: 'Output',
+            associatedClass: emptyClass,
+            isNullable: false,
+          ),
+          asynchronousType: AsynchronousType.callback,
+        )
       ])
     ], classes: <Class>[
       Class(name: 'Output', fields: <NamedType>[
@@ -2150,7 +2158,7 @@ void main() {
                     const TypeDeclaration(isNullable: false, baseName: 'int')),
           ],
           returnType: const TypeDeclaration(baseName: 'int', isNullable: false),
-          isAsynchronous: true,
+          asynchronousType: AsynchronousType.callback,
         )
       ])
     ], classes: <Class>[], enums: <Enum>[]);
@@ -2798,7 +2806,7 @@ void main() {
             associatedClass: emptyClass,
             isNullable: false,
           ),
-          isAsynchronous: true,
+          asynchronousType: AsynchronousType.callback,
         )
       ])
     ], classes: <Class>[
@@ -2897,7 +2905,7 @@ void main() {
         Method(
           name: 'doSomething',
           location: ApiLocation.flutter,
-          isAsynchronous: true,
+          asynchronousType: AsynchronousType.callback,
           parameters: <Parameter>[],
           returnType: TypeDeclaration(
             baseName: 'Enum1',
@@ -2940,7 +2948,7 @@ void main() {
         Method(
           name: 'doSomething',
           location: ApiLocation.flutter,
-          isAsynchronous: true,
+          asynchronousType: AsynchronousType.callback,
           parameters: <Parameter>[],
           returnType: TypeDeclaration(
             baseName: 'Enum1',
@@ -2988,7 +2996,7 @@ void main() {
         Method(
           name: 'doSomething',
           location: ApiLocation.host,
-          isAsynchronous: true,
+          asynchronousType: AsynchronousType.callback,
           parameters: <Parameter>[Parameter(name: 'value', type: enumType)],
           returnType: enumType,
         )
@@ -3032,7 +3040,7 @@ void main() {
         Method(
           name: 'doSomething',
           location: ApiLocation.host,
-          isAsynchronous: true,
+          asynchronousType: AsynchronousType.callback,
           parameters: <Parameter>[Parameter(name: 'value', type: enumType)],
           returnType: enumType,
         )