+++ packages/pigeon/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Added support of modern asynchronous api for Swift (async) and Kotlin (suspend).
 
 ## 22.7.0